Added test suit and Makefile
